requirements.txt
boto3==1.11.9
botocore==1.14.9
certifi==2019.11.28
chardet==3.0.4
Click==7.0
docutils==0.15.2
Flask==1.1.1
gunicorn==20.0.4
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
jmespath==0.9.4
MarkupSafe==1.1.1
numpy==1.18.1
picamera==1.13
pkg-resources==0.0.0
python-dateutil==2.8.1
python-dotenv==0.10.5
requests==2.22.0
s3transfer==0.3.2
six==1.14.0
urllib3==1.25.8
Werkzeug==0.16.0
